Fix build and refactor
