+++ python/mage/union_find/__init__.py
@@ -1,0 +1,4 @@
+from mage.union_find.union_find import (    # noqa: F401, F402, F403
+    UnionFind,
+    Node,
+)
