+++ python/tests/union_find/test_find.py
@@ -1,4 +1,4 @@
-from python.union_find import UnionFind
+from python.mage.union_find.union_find import UnionFind
 import pytest
 
 IDs = [i for i in range(10)]
@@ -11,12 +11,12 @@ def ds():
 
 class TestFind:
     def test_disconnected(self, ds):
-        assert ds.connected([0], [1]) == [False]
+        assert ds.connected(0, 1) is False
 
     def test_connected(self, ds):
-        ds.union([0], [1])
-        assert ds.connected([0], [1]) == [True]
+        ds.union(0, 1)
+        assert ds.connected(0, 1) is True
 
     def test_connected_later(self, ds):
         ds.union([0, 1], [1, 2])
-        assert ds.connected([0], [2]) == [True]
+        assert ds.connected(0, 2) is True
