+++ python/tests/union_find/test_init.py
@@ -1,4 +1,4 @@
-from python.union_find import UnionFind
+from python.mage.union_find.union_find import UnionFind
 import pytest
 
 IDs = [i for i in range(10)]
