+++ python/tests/union_find/test_union.py
@@ -1,4 +1,4 @@
-from python.union_find import UnionFind
+from python.mage.union_find.union_find import UnionFind
 import pytest
 
 IDs = [i for i in range(10)]
@@ -11,7 +11,7 @@ def ds():
 
 class TestUnion:
     def test_equal_height(self, ds):
-        ds.union([0], [1])
+        ds.union(0, 1)
         assert ds.nodes[1].parent == 0
 
     def test_different_height(self, ds):
